# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Phoenix Guardian API
  description: |
    Phoenix Guardian is an AI-powered healthcare security platform that provides
    real-time threat detection, secure transcription, and SOAP note generation
    for medical encounters.
    
    ## Authentication
    
    All API endpoints require authentication using JWT Bearer tokens.
    Tokens are obtained from the `/auth/token` endpoint and must include
    the tenant ID in the claims.
    
    ## Multi-Tenancy
    
    Phoenix Guardian is a multi-tenant platform. Each request must include
    the tenant context via the JWT token or X-Tenant-ID header.
    
    ## Rate Limiting
    
    - Standard: 1000 requests/minute per tenant
    - Transcription: 100 requests/minute per tenant
    - AI Operations: 50 requests/minute per tenant
    
    ## Versioning
    
    The API is versioned via the URL path (e.g., `/api/v1/`).
    
  version: 3.0.0
  contact:
    name: Phoenix Guardian API Support
    email: api-support@phoenix-guardian.health
    url: https://docs.phoenix-guardian.health
  license:
    name: Proprietary
    url: https://phoenix-guardian.health/license
servers:
  - url: https://api.phoenix-guardian.health/api/v1
    description: Production
  - url: https://staging-api.phoenix-guardian.health/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /encounters:
    get:
      summary: List encounters
      description: Retrieve paginated list of encounters for the tenant
      tags:
        - Encounters
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, cancelled]
        - name: physician_id
          in: query
          schema:
            type: string
            format: uuid
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of encounters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncounterList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
          
    post:
      summary: Create new encounter
      description: Start a new patient encounter
      tags:
        - Encounters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEncounter'
      responses:
        '201':
          description: Encounter created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Encounter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /threats/{threatId}:
    get:
      summary: Get threat details
      tags:
        - Threats
      parameters:
        - name: threatId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Threat details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Threat'
        '404':
          $ref: '#/components/responses/NotFound'
  /threats/{threatId}/acknowledge:
    post:
      summary: Acknowledge threat
      tags:
        - Threats
      parameters:
        - name: threatId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                assigned_to:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Threat acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Threat'
...
  /patients:
    get:
      summary: List patients
      tags:
        - Patients
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or MRN
      responses:
        '200':
          description: List of patients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientList'
                
    post:
      summary: Create patient
      tags:
        - Patients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatient'
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Encounter:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        patient_mrn:
          type: string
        physician_id:
          type: string
          format: uuid
        encounter_type:
          type: string
          enum: [office_visit, follow_up, consultation, urgent_care, telehealth]
        status:
          type: string
          enum: [active, completed, cancelled]
        chief_complaint:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        transcription_status:
          type: string
          enum: [pending, processing, completed, failed]
        soap_note_status:
          type: string
          enum: [pending, generated, reviewed, signed]
        language:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateEncounter:
      type: object
      properties:
        patient_mrn:
          type: string
        physician_id:
          type: string
          format: uuid
        encounter_type:
          type: string
          enum: [office_visit, follow_up, consultation, urgent_care, telehealth]
        chief_complaint:
          type: string
        language:
          type: string
          default: en
        start_time:
          type: string
          format: date-time
      required:
        - patient_mrn
        - encounter_type
...
    EncounterList:
      type: object
      properties:
        encounters:
          type: array
          items:
            $ref: '#/components/schemas/Encounter'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
          
    # Transcription
...
    Threat:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [prompt_injection, jailbreak, data_exfiltration, adversarial_audio, honeytoken_access, unauthorized_access]
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [detected, acknowledged, investigating, resolved]
        source:
          type: object
          properties:
            ip_address:
              type: string
            user_id:
              type: string
            encounter_id:
              type: string
        description:
          type: string
        details:
          type: object
        detected_at:
          type: string
          format: date-time
        acknowledged_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
        resolution:
          type: string
          
    # Patients
    Patient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        mrn:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other, unknown]
        preferred_language:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    CreatePatient:
      type: object
      properties:
        mrn:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other, unknown]
        preferred_language:
          type: string
          default: en
        phone:
          type: string
        email:
          type: string
          format: email
      required:
        - first_name
        - last_name
        - date_of_birth
    PatientList:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: '#/components/schemas/Patient'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
          
    # Dashboard
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object